Move `VCheckbox` into its own directory
